+++ madsim/src/sim/runtime/context.rs
@@ -7,8 +7,8 @@ use crate::{
 use std::{cell::RefCell, sync::Arc};
 
 thread_local! {
-    static CONTEXT: RefCell<Option<Handle>> = RefCell::new(None);
-    static TASK: RefCell<Option<Arc<TaskInfo>>> = RefCell::new(None);
+    static CONTEXT: RefCell<Option<Handle>> = const { RefCell::new(None) };
+    static TASK: RefCell<Option<Arc<TaskInfo>>> = const { RefCell::new(None) };
 }
 
 pub(crate) fn current<T>(map: impl FnOnce(&Handle) -> T) -> T {