+++ CHANGELOG.md
@@ -6,6 +6,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## tokio [0.2.26] - 2024-05-27
+
+### Added
+
+- Add a fake `runtime::Handle`. Note that this is a re-export of `madsim::task::Handle`. Tasks spawned with this handle are not correctly associated with the tokio runtime.
+
+## madsim [0.2.29] - 2024-05-27
+
+### Added
+
+- Implement `Debug` for `Handle` and `AbortHandle`.
+
+### Fixed
+
+- When `MADSIM_ALLOW_SYSTEM_THREAD` is set, executor won't panic when no task is ready.
+
 ## madsim [0.2.28] - 2024-05-17
 
 ### Added